Rename the loggingUtils module into logging
