+++ bin/distribution_test.py
@@ -30,7 +30,7 @@ from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession
 from fink_broker.distributionUtils import decode_kafka_df
 
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging import get_fink_logger, inspect_application
 
 
 def main():
