+++ bin/stream2raw.py
@@ -37,7 +37,7 @@ from fink_broker.parser import getargs
 from fink_broker.sparkUtils import from_avro
 from fink_broker.sparkUtils import init_sparksession, connect_to_kafka
 from fink_broker.sparkUtils import get_schemas_from_avro
-from fink_broker.loggingUtils import get_fink_logger, inspect_application
+from fink_broker.logging import get_fink_logger, inspect_application
 
 from fink_broker.monitoring import monitor_progress_webui
